backend/requirements.txt
mysql-connector-python
PyMySQL
mysqlclient
uvicorn
fastapi
sqlalchemy